Add dtach option in media many relation

+++ src/Lib/Media/Relations/MorphManyMedia.php
@@ -4,50 +4,42 @@ namespace Singsys\LQ\Lib\Media\Relations;
 
 use Singsys\LQ\Lib\Media\MediaUploader;
 use Illuminate\Database\Eloquent\Relations\MorphMany;
+use Illuminate\Database\Eloquent\Collection;
 
 class MorphManyMedia extends MorphMany {
 
+    use Concerns\MediaStoreUpdateRelation;
+
+    private $uploadedFiles = [];
+
     /**
-     * Store the Media File and update in relation.
+     * Store the file in default storage and update the relation in media table.
+     *
+     * @param array  $files     [Array Structure should be [[file => Blob, id => if already added]] ]
+     * @param string $path      [Destination path after Storage base path]
+     * @param boolean $detach   [false if don't want to remove previous uploaded files.]
+     * @param array $thumbnails [Thumbnails]
+     *
+     * @return $this
      */
-    public function addMedia(Array $files = null, $path = null, $thumbnails = null) {
+    public function addMedia(array $files = null, $path = null, $thumbnails = null, $detach = true) {
+        $this->uploadedFiles = new Collection();
+
         if ($files && !empty($files)) {
             $current_ids = [];
-            $uploaded_files = [];
             foreach ($files as $file) {
-                if (isset($file['file']) && !empty($file['file'])) {
-                    $uploader = new MediaUploader($file, $path, $thumbnails);
-                    $data = $uploader->uploadAndPrepareData();
-                    $data = array_merge($data, $this->make()->toArray());
-                    if (isset($file['id']) && !empty($file['id'])) {
-                        $media = clone $this->getQuery();
-                        $media = $media->where('id', $file['id'])->first();
-                        $media->update($data);
-                    } else {
-                        $media = $this->create($data);
-                    }
-
-                    $uploaded_files[] = $media;
+                $media = $this->_updloadFileUpdateRelation($file, $path, $thumbnails);
+                if ($media) {
+                    $this->uploadedFiles->push($media);
                     $current_ids[] = $media->id;
-
-                } else if (isset($file['id']) && !empty($file['id'])) {
-                    $media = clone $this->getQuery();
-                    $media = $media->where('id', $file['id'])->first();
-                    if ($media) {
-                        $uploaded_files[] = $media;
-                        $current_ids[] = $file['id'];
-                    }
                 }
             }
-
-            if (count($current_ids)) {
+            if (count($current_ids) && $detach) {
                 $unlinked = clone $this->getQuery();
                 $unlinked->whereNotIn('id', $current_ids)->delete();
             }
-
-            if($this->parent->mediaMorphRelation) {
-                $this->parent->setAttribute($this->parent->mediaMorphRelation, $uploaded_files);
-                $this->parent->setRelation($this->parent->mediaMorphRelation, $uploaded_files);
+            if ($this->parent->mediaMorphRelation) {
+                $this->parent->setRelation($this->parent->mediaMorphRelation, $this->uploadedFiles);
             }
 
         } else {
@@ -56,21 +48,7 @@ class MorphManyMedia extends MorphMany {
         }
         return $this;
     }
-
-    protected function unlinkRelation() {
-        # here we also need to delete the Media File.
-        $this->getQuery()->delete();
-    }
-
-    /**
-     * Get the class name of the parent model.
-     *
-     * @return string
-     */
-    protected function setMediaType($type)
-    {
-        $this->parent->setMediaMorphType($type);
-        $this->morphClass = $this->parent->getMorphClass();
-        return $this;
+    public function getMedia() {
+        return $this->uploadedFiles;
     }
 }
